requirements.txt
pytest
pandas
numpy
pyyaml
scipy
tensorboard
tabulate
gitpython
juliacall
CMake
bitsandbytes
tabulate
torcheval